SmartRecruiter/requirements.txt
streamlit
openai
pdfminer.six
pandas
requests
re
numpy
faiss-cpu
dotenv
ntlk
moviepy
soundfile